Migrated all collection components to functional components. Added rjsf types.
