+++ src/components/collection/CollectionAttributes.tsx
@@ -7,7 +7,7 @@ import type {
   CollectionData,
 } from "../../types";
 
-import React, { PureComponent } from "react";
+import React from "react";
 
 import * as BucketActions from "../../actions/bucket";
 import Spinner from "../Spinner";
@@ -31,17 +31,23 @@ export type Props = OwnProps &
     deleteCollection: typeof BucketActions.deleteCollection;
   };
 
-export default class CollectionAttributes extends PureComponent<Props> {
-  onSubmit = (formData: CollectionData) => {
-    const { match, updateCollection } = this.props;
+export default function CollectionAttributes({
+  match,
+  session,
+  bucket,
+  collection,
+  capabilities,
+  updateCollection,
+  deleteCollection,
+}: Props) {
+  const onSubmit = (formData: CollectionData) => {
     const {
       params: { bid, cid },
     } = match;
     updateCollection(bid, cid, { data: formData });
   };
 
-  deleteCollection = (cid: string) => {
-    const { deleteCollection, match } = this.props;
+  const handleDeleteCollection = (cid: string) => {
     const {
       params: { bid },
     } = match;
@@ -54,42 +60,41 @@ export default class CollectionAttributes extends PureComponent<Props> {
     }
   };
 
-  render() {
-    const { match, session, bucket, collection, capabilities } = this.props;
-    const {
-      params: { bid, cid },
-    } = match;
-    const { busy, data: formData } = collection;
-    if (busy) {
-      return <Spinner />;
-    }
-    return (
-      <div>
-        <h1>
-          Edit{" "}
-          <b>
-            {bid}/{cid}
-          </b>{" "}
-          collection attributes
-        </h1>
-        <CollectionTabs
+  const {
+    params: { bid, cid },
+  } = match;
+  const { busy, data: formData } = collection;
+
+  if (busy) {
+    return <Spinner />;
+  }
+
+  return (
+    <div>
+      <h1>
+        Edit{" "}
+        <b>
+          {bid}/{cid}
+        </b>{" "}
+        collection attributes
+      </h1>
+      <CollectionTabs
+        bid={bid}
+        cid={cid}
+        selected="attributes"
+        capabilities={capabilities}
+      >
+        <CollectionForm
           bid={bid}
           cid={cid}
-          selected="attributes"
-          capabilities={capabilities}
-        >
-          <CollectionForm
-            bid={bid}
-            cid={cid}
-            session={session}
-            bucket={bucket}
-            collection={collection}
-            deleteCollection={this.deleteCollection}
-            formData={formData}
-            onSubmit={this.onSubmit}
-          />
-        </CollectionTabs>
-      </div>
-    );
-  }
+          session={session}
+          bucket={bucket}
+          collection={collection}
+          deleteCollection={handleDeleteCollection}
+          formData={formData}
+          onSubmit={onSubmit}
+        />
+      </CollectionTabs>
+    </div>
+  );
 }
