+++ src/components/collection/CollectionCreate.tsx
@@ -5,7 +5,7 @@ import type {
   BucketRouteMatch,
 } from "../../types";
 
-import React, { PureComponent } from "react";
+import React from "react";
 
 import * as BucketActions from "../../actions/bucket";
 import Spinner from "../Spinner";
@@ -26,32 +26,37 @@ export type Props = OwnProps &
     createCollection: typeof BucketActions.createCollection;
   };
 
-export default class CollectionCreate extends PureComponent<Props> {
-  render() {
-    const { match, session, bucket, collection, createCollection } = this.props;
-    const {
-      params: { bid },
-    } = match;
-    const { busy } = session;
-    if (busy) {
-      return <Spinner />;
-    }
-    return (
-      <div>
-        <h1>
-          Create a new collection in <b>{bid}</b> bucket
-        </h1>
-        <div className="card">
-          <div className="card-body">
-            <CollectionForm
-              session={session}
-              bucket={bucket}
-              collection={collection}
-              onSubmit={formData => createCollection(bid, formData)}
-            />
-          </div>
+export default function CollectionCreate({
+  match,
+  session,
+  bucket,
+  collection,
+  createCollection,
+}: Props) {
+  const {
+    params: { bid },
+  } = match;
+  const { busy } = session;
+
+  if (busy) {
+    return <Spinner />;
+  }
+
+  return (
+    <div>
+      <h1>
+        Create a new collection in <b>{bid}</b> bucket
+      </h1>
+      <div className="card">
+        <div className="card-body">
+          <CollectionForm
+            session={session}
+            bucket={bucket}
+            collection={collection}
+            onSubmit={formData => createCollection(bid, formData)}
+          />
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
